requirements.txt
pytest==7.4.3
aws-cdk-lib==2.111.0
constructs>=10.0.0,<11.0.0
pycodestyle==2.11.1
